chore: move toolbar/items to toolbar/desktop/items
