+++ lib/src/editor/toolbar/desktop/desktop.dart
@@ -1,0 +1,14 @@
+export 'floating_toolbar.dart';
+export 'floating_toolbar_widget.dart';
+
+export 'items/bulleted_list_toolbar_item.dart';
+export 'items/color/color.dart';
+export 'items/format_toolbar_items.dart';
+export 'items/heading_toolbar_items.dart';
+export 'items/highlight_toolbar_item.dart';
+export 'items/icon_item_widget.dart';
+export 'items/link/link_toolbar_item.dart';
+export 'items/numbered_list_toolbar_item.dart';
+export 'items/paragraph_toolbar_item.dart';
+export 'items/placeholder_toolbar_item.dart';
+export 'items/quote_toolbar_item.dart';
