+++ lib/src/editor/toolbar/desktop/items/format_toolbar_items.dart
@@ -1,5 +1,5 @@
 import 'package:appflowy_editor/appflowy_editor.dart';
-import 'package:appflowy_editor/src/editor/toolbar/items/utils/tooltip_util.dart';
+import 'package:appflowy_editor/src/editor/toolbar/desktop/items/utils/tooltip_util.dart';
 
 final List<ToolbarItem> markdownFormatItems = [
   _FormatToolbarItem(
