+++ lib/src/editor/toolbar/desktop/items/link/link_toolbar_item.dart
@@ -1,5 +1,5 @@
 import 'package:appflowy_editor/appflowy_editor.dart';
-import 'package:appflowy_editor/src/editor/toolbar/items/link/link_menu.dart';
+import 'package:appflowy_editor/src/editor/toolbar/desktop/items/link/link_menu.dart';
 import 'package:appflowy_editor/src/infra/clipboard.dart';
 import 'package:flutter/material.dart';
