+++ lib/src/editor/toolbar/desktop/items/utils/overlay_util.dart
