+++ lib/src/editor/toolbar/toolbar.dart
@@ -1,19 +1,3 @@
-export 'desktop/floating_toolbar.dart';
-export 'desktop/floating_toolbar_widget.dart';
-
-export 'items/heading_toolbar_items.dart';
-export 'items/paragraph_toolbar_item.dart';
-export 'items/placeholder_toolbar_item.dart';
-export 'items/format_toolbar_items.dart';
-
-export 'items/bulleted_list_toolbar_item.dart';
-export 'items/numbered_list_toolbar_item.dart';
-export 'items/quote_toolbar_item.dart';
-
-export 'items/link/link_toolbar_item.dart';
-export 'items/color/color.dart';
-export 'items/highlight_toolbar_item.dart';
-
-export 'items/icon_item_widget.dart';
+export 'desktop/desktop.dart';
 
 export 'mobile/mobile.dart';
