+++ test/new/toolbar/desktop/items/color/color_picker_test.dart
@@ -1,8 +1,8 @@
-import 'package:appflowy_editor/src/editor/toolbar/items/color/color_picker.dart';
+import 'package:appflowy_editor/src/editor/toolbar/desktop/items/color/color_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../../../infra/testable_editor.dart';
+import '../../../../infra/testable_editor.dart';
 
 void main() {
   setUpAll(() {
